+++ spark/src/test/scala_2.13-/org/datatools/bigdatatypes/spark/SqlTypeConversionSparkSpec.scala
@@ -3,11 +3,13 @@ package org.datatools.bigdatatypes.spark
 import org.apache.spark.sql.types.{IntegerType, StringType, StructField, StructType}
 import org.datatools.bigdatatypes.TestTypes._
 import org.datatools.bigdatatypes.UnitSpec
-import org.datatools.bigdatatypes.conversions.SqlTypeConversion
+import org.datatools.bigdatatypes.conversions.{SqlInstanceConversion, SqlTypeConversion}
 import org.datatools.bigdatatypes.formats.Formats.implicitDefaultFormats
 import org.datatools.bigdatatypes.spark.SqlTypeConversionSpark._
 import org.datatools.bigdatatypes.types.basic.{Nullable, Required, SqlInt, SqlString, SqlStruct, SqlType}
 
+/** Reverse conversion, from Spark types to [[SqlType]]s
+  */
 class SqlTypeConversionSparkSpec extends UnitSpec {
 
   "Simple Spark DataType" should "be converted into SqlType" in {
@@ -17,13 +19,13 @@ class SqlTypeConversionSparkSpec extends UnitSpec {
   "StructField nullable" should "be converted into Nullable SqlType" in {
     val sf = StructField("myInt", IntegerType, nullable = true)
     sf.getType shouldBe SqlInt(Nullable)
-    SqlTypeConversionSpark(sf).getType shouldBe SqlInt(Nullable)
+    SqlInstanceConversion[StructField].getType(sf) shouldBe SqlInt(Nullable)
   }
 
   "StructField required" should "be converted into Required SqlType" in {
     val sf = StructField("myInt", IntegerType, nullable = false)
     sf.getType shouldBe SqlInt()
-    SqlTypeConversionSpark(sf).getType shouldBe SqlInt(Required)
+    SqlInstanceConversion[StructField].getType(sf) shouldBe SqlInt(Required)
   }
 
   "StructType" should "be converted into SqlStruct" in {
@@ -31,7 +33,7 @@ class SqlTypeConversionSparkSpec extends UnitSpec {
     val sf2 = StructField("myString", StringType, nullable = true)
     val st = StructType(List(sf, sf2))
     val expected = SqlStruct(List(("myInt", SqlInt()), ("myString", SqlString(Nullable))))
-    SqlTypeConversionSpark(st).getType shouldBe expected
+    SqlInstanceConversion[StructType].getType(st) shouldBe expected
     st.getType shouldBe expected
   }
 
@@ -76,6 +78,5 @@ class SqlTypeConversionSparkSpec extends UnitSpec {
     val sqlType: SqlType = SparkTypes[ExtendedTypes].sparkSchema.getType
     sqlType shouldBe extendedTypes
   }
-
 
 }